【PIR】Improve some operations to support more scenarios

+++ tests/test_auto_scan_shape.py
@@ -28,7 +28,7 @@ class Net(BaseNet):
         """
         forward
         """
-        x = paddle.shape(inputs)
+        x = paddle.shape(inputs).astype("int32")
         return x
